pyproject.toml
[tool.poetry]
name = "cover-letter"
version = "0.1.0"
description = ""
authors = ["Swarnashruti Jupudi <swarna.s92@gmail.com>"]
readme = "README.md"
packages = [{include = "cover_letter"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.3.3"
pydantic = "^2.5.1"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
